requirements.txt
matplotlib>=3.9.3
numpy>=2.0.2
optuna>=3.6.1
Pillow>=10.0.0
pygame>=2.6.1
gymnasium>=0.29.1
pettingzoo[mpe]>=1.24.3
pytest>=8.0.0
torch>=2.6.0
